Update liu2024.py
When comparing the real trial events with the encoded events, they are inverted: in the real data, 1 corresponds to encoded 2, and in the real data, 2 corresponds to encoded 1. To clarify the distinction between the right hand (1) and left hand (2) in both the real and encoded data, the mapping rules and encoded_mapping have been adjusted, resulting in a more accurate alignment of the test types from the source.

Signed-off-by: ChristianFonsecaRodriguez <125149800+ChristianFonsecaRodriguez@users.noreply.github.com>